refactor(plugin-cursor): split discover tests by domain (M7)
Nine discover-*.test.ts modules + test-support/discover-helpers.ts replace monolithic
discover.test.ts; 66 discover cases, 137 package tests unchanged. Memory bank: L2 build
complete for slobac-audit-remediation-m7.
