+++ memory-bank/active/activeContext.md
@@ -2,14 +2,13 @@
 
 ## Current Task: SLOBAC Audit Remediation — M7 (Split plugin-cursor discover.test.ts)
 
-**Phase:** L2 PREFLIGHT — COMPLETE (ready for build)
+**Phase:** L2 BUILD — COMPLETE (ready for `/niko-qa`)
 
 ## What Was Done
 
-- `/niko` Step 2a: M6 checked off; L4 sub-run ephemeral state cleared; M7 classified Level 2.
-- **Plan:** Full `tasks.md` — four `discover-*.test.ts` modules + `discover-helpers.ts`; parity 63 / 137.
-- **Preflight:** PASS — recorded below; `.preflight-status` written.
+- **Build:** Nine `discover-*.test.ts` modules + `test-support/discover-helpers.ts`; deleted `discover.test.ts`. Paths: `packages/plugin-cursor/test/discover-cursor-plugin.test.ts`, `discover-mdc-parsing.test.ts`, `discover-file-rule.test.ts`, `discover-simple-agent-skill-rules.test.ts`, `discover-classification-priority.test.ts`, `discover-agent-ignore.test.ts`, `discover-skills.test.ts`, `discover-commands.test.ts`, `discover-agent-skill-io.test.ts`, `test-support/discover-helpers.ts`.
+- **Verification:** `pnpm test` at repo root green; `@a16njs/plugin-cursor` 137 tests (66 discover + 71 other).
 
 ## Next Step
 
-**Build:** Load `.cursor/skills/shared/niko/references/level2/level2-build.md` and execute the implementation plan in `tasks.md`.
+Run **`/niko-qa`** for L2 semantic review, then **`/niko-reflect`**, then **`/niko`** for Step 2a (check off M7 in `milestones.md`).
